Moving sidebar heading components to shared filed
