+++ src/components/sidebar/components/sidebar-headings/sidebar-headings.module.css
@@ -1,0 +1,10 @@
+.root {
+  color: var(--token-color-foreground-primary);
+  font-size: var(--token-typography-display-100-font-size);
+  font-weight: var(--token-typography-font-weight-semibold);
+  line-height: var(--token-typography-display-100-line-height);
+  margin-bottom: 12px;
+  margin-top: 0;
+  padding-left: 8px;
+  padding-right: 8px;
+}
