+++ src/components/sidebar/components/sidebar-headings/types.ts
@@ -1,0 +1,10 @@
+interface SidebarSectionHeadingProps {
+  text: string
+}
+
+interface SidebarTitleHeadingProps {
+  id: string
+  text: string
+}
+
+export type { SidebarSectionHeadingProps, SidebarTitleHeadingProps }
